// src/app/models/admin/capphephoatdongkhoangsan/cptanthukhoangsan/cptanthudvhc.model.ts
export class OutputCpTanThuDvhcModel {
  idtanthudvhc: string;
  idxa: string;
  idhuyen: string;
  idtinh: string;
  maxa: string;
  mahuyen: string;
  matinh: string;
  idcappheptanthu: string;
  tentinh: string;
  tenhuyen: string;
  tenxa: string;
  siteid: string;
}

export class InputCpTanThuDvhcModel {
  idtanthudvhc: string;
  idxa: string;
  idhuyen: string;
  idtinh: string;
  maxa: string;
  mahuyen: string;
  matinh: string;
  idcappheptanthu: string;
  siteid: string;
}